requirements.txt
asgiref==3.7.2
colorama==0.4.6
crispy-bootstrap5==2023.10
distlib==0.3.8
Django==5.0.2
django-crispy-forms==2.1
django-widget-tweaks==1.5.0
filelock==3.13.1
pillow==10.2.0
platformdirs==4.2.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
virtualenv==20.25.0
